Unwrap nullability-wrapped symbols in object initializer completion
This slipped through because we didn't have any tests at all for
the case where we are using the type-inferrer to infer the case like:

    a = new A { B = { } }

where the inner one doesn't have a new B before it. The scenario wasn't
new but just happened to be the precise case that was broken.

Fixes https://github.com/dotnet/roslyn/issues/36702