Don't lose tree DiganosticOptions when we recover or replace a tree
We were not passing through DiagnosticOptions when we created a
recoverable tree or when we called Solution.WithDocumentSyntaxRoot. This
implementation takes a straightforward approach that we simply hold
onto the DiagnosticOptions in the process (rather than recovering it);
we could improve on that but it makes things a bit trickier since
DiagnosticOptions would have to be fetched on demand.

Fixes https://github.com/dotnet/roslyn/issues/26391