Workaround IExperimentationService having a UI thread dependency
Right now we are seeing deadlocks where MEF is trying to create
the RoslynVisualStudioWorkspace on a background thread, and that
deadlocks because the IExperimentationService it tries to use has a
UI thread dependency. This breaks that specific cycle by avoiding using
the IExperimentationService until on the UI thread -- we know we can't
be asked for document options until we have documents, and that can't
happen until we get projects. Thankfully, the creation of projects
are still UI affinitized so we can initialize it there.