Switch from FirstOrDefault to First
This code is assuming that given a token, we can chase up and find
a node that surrounds a span. Using FirstOrDefault meant if we couldn't
find one we'd pass null to GetSemanticModelForNodeAsync which would
promptly throw an exception. This will still throw exceptions but at
least will throw earlier.