Fix using statements to use correct copy of PooledObjects
The Microsoft.CodeAnalysis.FlowAnalysis.Utilities assembly which we
consume accidentally made it's copy of the PooledObjects helper public.
That copy has a different namespace, so by using the wrong namespace
we were pulling in that assembly's copy rather than using the copy we
have in our assembly. This causes us to do an early assembly load when
it's absolutely not necessary.