Fix the type inferrer's triggering of method type inferrence
The type inferrer does method type inferrence if we know what the
return type of the method should be. This triggered if the method was
handled an argument which it could then chase up and find the
InvocationExpressionSyntax. Unfortunately if you are inferring and
don't have an argument you still want to be able to infer the type
of the argument if there was one there, but there's no ArgumentSyntax
to work from. Thus, it's easier to just directly pass the
InvocationExpressionSyntax and avoid the tree walk; looking through
the callers of InferTypeInArgument there's only one place where
it could potentially find an InvocationExpressionSyntax, and that one
place is the place I updated the call site on.