Convert the type inferrer tests to be Theories
Most of the type inferrer tests test both overloads of the
type inferrence service -- the one that takes a span and one that
takes a position. Both (by default) were tested, but it's impossible
without actually debugging a test to know which one failed. I had a
particular irritating case where I made a behavior change and had to
update a bunch of tests with new baselines, but since the behavior
also split between both APIs it was confusing to figure out what was
going on.