Move call to RegisterDocumentOptionsProvider to TestWorkspace constructor
This code as was written was a bit shaky. The Workspace constructor
would ask for the workspace option service, and this would cause us
(while creating the service) to call RegisterDocumentOptionsProvider.
If that results in one of the options providers also asking for options,
we recursively request the option service while still trying to
construct it.

The fix is just to move the registration to the workspace constructor
itself which removes the circularity and matches the pattern used
everywhere else.