Revert underselection prevention in RefactoringHelperService TryGetRelevantNodeAsync.
The underselection is desirable in a number of scenarios:
- only header is selected
- only identifier (name of function) is selected
It would be possible to special case them but that's little benefit in
comparison to added complexity.

The version that is being reverted is incorrect even despite
abovementioned things. It doesn't take into account extractedNode x
selectedNode (for it to work it the underselection would need to be
tested against both).

- ...
- there's underselection