Remove some duplicate helpers around removing attributes
It turns out for methods we already had
RemoveInaccessibleAttributesAndAttributesOfTypes but had accidentally
created another helper for filtering out attributes from parameters
that more correctly should have used that. This also observed that we
didn't have a helper for IParameterSymbols to do the same thing.
We did have one that filtered but not one that removed innaccessible
parameters, which was actually a bug.