Update code action test helpers to dump diagnostics if there's no fix
If you write a test to assert a code fix works, and you don't get a
code fix, sometimes that's because the code fix is broken. Sometimes
it's just because you have a silly typo in your test that means the
test workspace isn't doing what you think. To aid in debugging those,
add some extra test support to dump diagnostics in that case.