Do not fade code for IDE0016 (UseThrowExpressionDiagnosticId)
Fixes #27213

Analyzer recommends using throw expression for null checks instead of throw statement within an if statement with null check. We used to fade the entire if statement for discoverability and increased code fix span, but this was undesirable as we want to fade only unnecessary or dead code. This change removes the fading diagnostic and updates unit tests accordingly