Delete protected constructor and property setter
This constructor never should have been there, as it allows the
potential public creation of a type that won't have assigned any of
it's fields required for the type to work. This is technically a
breaking change, but there's no reasonable way somebody could have
used this in a functional scenario.

We also had a property setter that shouldn't have been there, but
can now be removed.