Restore constraint checking for reduced extension methods. (#38109)
Before C# 8, the reduction of an extension method would fail (return null) when the
method's constraints would not be satisfied.  But in C# 8 the unmanaged constraint requires
we know the language version, and therefore the compilation, to check.  That is not available
in `IMethodSymbol.ReduceExtensionMethod`.  So we dropped the constraint checking in #37022.
However this causes an unfortunate regression in the quality of the IDE experience.  So here we
restore the constraint checking, using the compilation's language version when available.
IMethodSymbol.ReduceExtensionMethod uses the C# 8 rules since it has no language version to check.

Fixes #37780