Call GetTextSynchronously when fetching in a synchronous way
This code is trying to fetch the text of a single line, but since it's
responding to an API that's synchronous it must be synchronous.
We were doing it with GetTextAsync + WaitAndGetResult which can be slow
if the thread pool is busy.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/974884