Remove TestableDesktopBuildClient
I was able to collapse this into the base definition of
`DesktopBuildClient`.

This change did remove one test which was adding unnecessary complexity.
The test was verifying the exception behavior of an internal function.
This behavior is mostly covered by other tests and the fine grained
exception hook wasn't worth the other complexity to the code base.