Disable failing tests
Related to #38454.

These tests fail in the CI runs for #38417 which adds translations for various resources. The problem is the test compares localized text with the original English text; this worked fine until now because the localized text wasn't yet translated, and so just happened to be the same as the English.

The point of this change is to unblock #38417.