Redo when we create text buffers in TestWorkspace
With TestWorkspace we were creating text buffers for all types of
documents up front, and by default opening all buffers in the workspace.
This was problematic because if we're creating text buffers we should
be ensuring the text buffer's options for using tabs and spaces matches
the Workspace options, but the workspace sometimes wasn't available
at those times. This is also unnecessarily expensive if you have test
projects that aren't doing anything with editor experiences at all.

Also, a complicated "create a separate document in the TestWorkspace for
the cursor" was removed in favor of more direct code that just creates
a buffer. Creating an entire TestDocument was funky since the document
was never actually part of the workspace -- this revealed that it didn't
seem we were ever disposing the views created.

The "languageName" parameter to CreateProjectionBuffer was removed: it
didn't logically make sense (the projection buffer created isn't in
a Workspace, so it has no language name), but also wasn't being used
for anything else.