Fix crash where we are not always finding text snapshots
Given this is not a reliable crash, my guess is the problem is we are
sometimes processing an older Solution snapshot. Asking a specific
SourceText for the specific ITextSnapshot it came from could fail
if that snapshot had been GC'ed, whereas going to the container
first and asking for the buffer directly should be more reliable. Either
way we now also handle the null case.

Fixes https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1010045