Temporarily publish to only one channel, with Shipping packages only (#39742)
This is causing a race where we try to submit multiple changes to the dotnet/versions repository one after the other. The changes need to be fast-forward, but if you submit changes very quickly, you could get the HEAD of the old build right after submitting, meaning that the commit would be based on the old content. The failure to commit a fast-forward merge is failing the build.

dotnet/arcade#4320 addresses the need to retry if a fast-forward fails.