Implement a method to get the spacing for the closing brace
The closing brace is calculated based on what kind of syntax we
are dealing with. If it is a DeclarationKind then we can just align
it based on the span start. If it is a return statement we need to
add an extra tab for spacing because there is no declarationkind for
returns and we passed our starting point.