EnC: Fix document out-of-sync checks when module is not loaded when d… (#39836)
* EnC: Fix document out-of-sync checks when module is not loaded when debugging session starts

Previous change https://github.com/dotnet/roslyn/pull/39295 allowed us to rely on PDB source file checksums when distinguishing between design-time-only documents that should be ignored when applying changes and documents whose changes must be applied to the debuggee. The change did not handle the case when the PDB isn't loaded yet when this check is being performed. This bug resulted in changes not being applied correctly and documents getting out-of-sync in scenarios where the modules were not loaded to the debuggee fast enough.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1013242