Remove Live Share's use of a ProjectTracker and AbstractProject
Live Share was still using a bunch of deprecated shims which in some
cases were only existing to support this code. Delete it all and move
onto direct uses of the Workspaces API.

There appears to be a memory leak in the existing code: it doesn't
appear the projects that were created for Razor contained documents were
ever removed from the workspace. This maintains that leak since it's not
obvious how to fix it, but at least the leak is more obvious now.