Add a DeclarationKind that maps to the return syntax kinds
This allows us to walk up nodes and determine what kind of declaration
there is. The return declaration is needed for the Collection Initializer
Wrappers. We need to know when we have reached the beginning of the line
to make sure the opening brace is indented properly.

All other declaration kinds were available except the return. This way we
know when we have reached the beginning of the line. Before, we would pass
the return statement and end up at the wrong indentation level.