Fix possible null reference
In the deepest of ironies, the extension methods we had to wrap
top-level nullability was implicitly hiding a potential null reference.
Removing the extension methods is now flagging this, so this fixes it
up. The handling isn't ideal: if we don't have a Task<T> we won't
necessarily wrap, but that's not something really worth going out of
our way to support -- most code doesn't.