Unskip tests that were skipped due to adaptive formatting triggering
We had tests that set "use tabs" but then were working in a file of
mostly spaces. The editor behavior now is to treat this as if the
file had the "use spaces" on. I alternatively could have chosen to
disable inferrence, but then it seems we're not actually testing the
more mainline scenario of somebody having tabs in the entire file, which
seems like it's worth testing over odd mixed cases.

Closes https://github.com/dotnet/roslyn/issues/39967