Remove CompilationWithAnalyzers.IsDiagnosticAnalyzerSuppressed checks in the IDE Diagnostic service
Disabled by default analyzers, which are enabled via .editorconfig are not executing in the IDE diagnostic service. This is due to us skipping live analysis for analyzers disabled by default OR disabled via compilation options as a performance optimization. However, this check is no longer correct as analyzers can be enabled/disabled for individual documents through .editorconfig files. So we no longer perform this check and pessimistically assume analyzer is not suppressed and let the core analyzer driver in the compiler layer handle skipping redundant analysis callbacks.