Remove null from Diagnostic.Arguments
This changes the implementation of `DiagnosticInfo` to use an empty
array in all cases for arguments vs. a mix of empty arrays and `null`
for the empty case. This allows us to ensure the `Diagnostic.Arguments`
is always a non-null value.

This makes adopting nullability higher in the stack significantly
simpler as there are assumptions this is non-null in places.