Add support for a --compiler-invocation switch to create Compilations
For some build systems or environments, rather than having the LSIF
tool run MSBuildWorkspace we'd rather it just reuse an existing
compiler invocation that was previously logged. For example, if you know
the command line string that was passed to the compiler as a part of a
build, you can do the LSIF analysis later without having to run MSBuild.
This is handy for non-MSBuild-based build systems, or in some
distributed build systems which supports the concept of doing analyzer-
like things on different machines in parallel with the main build.

This adds a new switch, --compiler-invocation that points to a JSON
file that contains:

- Command line switches.
- The path to the project file
- Path mapping information

The latter is useful if you are trying to run this analysis on a
different machine than the original build happened on; for example, if
the command line switches were on a machine where your source was under
S:\Source, and now your machine has the same sources, but under
S:\Source2, the JSON file can contain the original command line string
but we'll map everything when producing the compilation.