Move to target netcoreapp3.1
Change Roslyn to target netcoreapp3.1 when building .NET Core assets.

Previously the code targetted a mix of netcoreapp2.1 and netcoreapp3.0.
The mix is due to default interfaces only being supported on
netcoreapp3.0 and hence our testing needed to use that. Yet at the same
time we were required to ship the compiler in SDKS that targetted
netcoreapp2.1. Now we can universally target netcoreapp3.1.