Remove fourth state of SyntaxNodeOrToken
The `SyntaxNodeOrToken` type had four states:

1. It's a `SyntaxNode`
1. It's a `SyntaxToken`
1. It's `default(SyntaxNodeOrToken)`
1. It returns `true` for `IsNode` but `AsNode` returns `null`

The fourth state appears to be an accident of implementation. This
change removes that state. The one place in the code where this appeared
to be an issue is in the implicit conversion from `SyntaxNode?` to
`SyntaxNodeOrToken`. In all places where this was used and the value was
potentially `null` the caller only used `GetLocation` on the result.
Hence the fix is simply to change that API to return `default` instead
of the fourth state in the case of `null`