Fix MS.VB reference
The TPA now considers the output directory as a trusted place for
loading assemblies. To the point it will prefer certain DLLs in the
output directory over equal ones that ship with the runtime.

This broke a number of VB tests here because we unconditionally
referenced the desktop version of Microsoft.VisualBasic. This got copied
to the output directory, included in the TPA and hence loaded during
test execution. This breaks tests because they require the .NET Core
version of Microsoft.VisualBasic.

Conditioned the reference to be desktop only so it's not present on our
.NET Core builds