Delete the lazy caching of GetLatestProjectVersion()
Solution.GetLatestProjectVersion() returns the highest project version
of any of the Projects in the Solution. This is lazily computed, but
the lazy computation has a fatal flaw. It's implemented with a
Lazy<VersionStamp>, and if a Solution is changed and no project was
changed (say, you only modified a document's text), the existing
Lazy<VersionStamp> is reused. This causes us to root an older
SolutionState or ProjectState until somebody were to call
GetLatestProjectVersion() and cause the Lazy<VersionStamp> to be
evaluated.

Except...that would never happen. It turns out that nothing calls
GetLatestProjectVersion() in Visual Studio anywhere that I can find.
The only call is in one unit test that calls it and then ignores the
result. Thus, we'd simply leak the SolutionState/ProjectState until
some code ran that knew it could explicitly state the value instead of
reuse the existing Lazy<VersionStamp>.

Looking through history, it appears that this used to be used more
regularly as a part of an older implementation of the
ProjectDependencyGraph, but that was removed in
abd7654e3b878cf6a2852e62bf6055a755fb6965.

Since nobody calls this, and the actual computation (just looping
through all projects to find the newest stamp) is fairly trivial, I'm
just going to delete all the lazy handling and just inline the
computation. In the unlikely case that this ever becomes a hot path,
we can revisit the laziness at that time.